Busy: move scene definitions file
